chore: :recycle: refactor to put commands & functions into own files
